refactor : user 객체 사용하던 클래스들 전부 수정
