+++ src/main/java/com/kdev5/cleanpick/customer/domain/PreferredManagerId.java
@@ -14,7 +14,7 @@ public class PreferredManagerId implements Serializable {
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (!(o instanceof com.kdev5.cleanpick.user.domain.PreferredManagerId that)) return false;
+        if (!(o instanceof com.kdev5.cleanpick.customer.domain.PreferredManagerId that)) return false;
         return Objects.equals(customerId, that.customerId) &&
                 Objects.equals(managerId, that.managerId);
     }
