+++ src/main/java/com/kdev5/cleanpick/manager/domain/ManagerAvailableTime.java
@@ -1,7 +1,6 @@
 package com.kdev5.cleanpick.manager.domain;
 
 import com.kdev5.cleanpick.global.entity.BaseTimeEntity;
-import com.kdev5.cleanpick.user.domain.User;
 import jakarta.persistence.*;
 
 import java.time.LocalTime;
@@ -15,7 +14,7 @@ public class ManagerAvailableTime extends BaseTimeEntity {
 
     @ManyToOne
     @JoinColumn(name = "manager_id", nullable = false)
-    private User manager;
+    private Manager manager;
 
     @Column(name = "day_of_week", nullable = false)
     private int dayOfWeek;
