+++ src/main/java/com/kdev5/cleanpick/payment/domain/PaymentMethod.java
@@ -1,9 +1,9 @@
 package com.kdev5.cleanpick.payment.domain;
 
+import com.kdev5.cleanpick.customer.domain.Customer;
 import com.kdev5.cleanpick.global.entity.BaseTimeEntity;
 import com.kdev5.cleanpick.payment.domain.enumeration.PayType;
 import jakarta.persistence.*;
-import com.kdev5.cleanpick.user.domain.User;
 
 @Entity
 @Table(name = "payment_method")
@@ -13,8 +13,8 @@ public class PaymentMethod extends BaseTimeEntity {
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
+    @JoinColumn(name = "customer_id", nullable = false)
+    private Customer customer;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
